backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.12
motor==3.1.1
requests==2.31.0
pytest==7.4.0
pytest-asyncio==0.22.0
httpx==0.24.1
python-dotenv==1.0.0
openai==1.32.0
coverage==7.4.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
PyPDF2==3.0.1
